use SDCFlows brain extraction workflow for T2w when T1w has precomputed mask

+++ nibabies/workflows/anatomical/outputs.py
@@ -376,7 +376,7 @@ def init_anat_derivatives_wf(
     )
 
     ds_t2w_preproc = pe.Node(
-        DerivativesDataSink(base_directory=output_dir, space="T1w", desc="preproc", compress=True),
+        DerivativesDataSink(data_dtype="i2", base_directory=output_dir, space="T1w", desc="preproc", compress=True),
         name="ds_t2w_preproc",
         run_without_submitting=True,
     )
